refactor(skills): split watcher target construction
Keep refresh runtime within the enforced line limit without adding a new suppression by extracting self-contained watch-target construction.
